Struct dslab_iaas::core::config::SimulationConfigRaw

pub struct SimulationConfigRaw {

    pub send_stats_period: Option<f64>,
    pub message_delay: Option<f64>,
    pub allocation_retry_period: Option<f64>,
    pub vm_start_duration: Option<f64>,
    pub vm_stop_duration: Option<f64>,
    pub allow_vm_overcommit: Option<bool>,
    pub network_throughput: Option<u64>,
    pub simulation_length: Option<f64>,
    pub number_of_hosts: Option<u32>,
    pub host_cpu_capacity: Option<f64>,
    pub host_memory_capacity: Option<f64>,
    pub step_duration: Option<f64>,
    pub vm_allocation_timeout: Option<f64>,
    pub hosts: Option<Vec<HostConfig>>,
    pub schedulers: Option<Vec<SchedulerConfig>>,

}

Fields
send_stats_period: Option<f64>

periodically send statistics from host to monitoring

message_delay: Option<f64>

message trip time from any host to any direction

allocation_retry_period: Option<f64>

when allocation request fails then wait for this duration

vm_start_duration: Option<f64>

vm initialization duration

vm_stop_duration: Option<f64>

vm deallocation duration

allow_vm_overcommit: Option<bool>

pack VM by real resource consumption, not SLA

network_throughput: Option<u64>

currently used to define VM migration duration

simulation_length: Option<f64>

length of simulation (for public datasets only)

number_of_hosts: Option<u32>

number of hosts in datacenter (for public datasets only)

host_cpu_capacity: Option<f64>

CPU capacity for default host

host_memory_capacity: Option<f64>

RAM capacity for default host

step_duration: Option<f64>

duration beetween user access the simulation info

vm_allocation_timeout: Option<f64>

VM becomes failed after this timeout is reached

hosts: Option<Vec<HostConfig>>

cloud physical hosts

schedulers: Option<Vec<SchedulerConfig>>

cloud schedulers