Module dslab_iaas::core::host_manager

Host manager representing a physical machine.

Structs

HostManager

Represents a single physical machine or host for short, which possesses a certain amount of resources and performs execution of VMs assigned to it by a scheduler. It models the main VM lifecycle stages such as creation, deletion and migration, and reports the VM status changes to VM API component. Host manager periodically computes its current load, as the sum of loads produced by currently running VMs, and reports it to the monitoring component. Host manager also records the total energy consumption of the host computed using the power model defined as a function of CPU load.

SendHostState