Struct dslab_iaas::core::host_manager::HostManager

pub struct HostManager {
    pub id: u32,
    pub rack_id: Option<u32>,
    /* private fields */
}

Represents a single physical machine or host for short, which possesses a certain amount of resources and performs execution of VMs assigned to it by a scheduler. It models the main VM lifecycle stages such as creation, deletion and migration, and reports the VM status changes to VM API component. Host manager periodically computes its current load, as the sum of loads produced by currently running VMs, and reports it to the monitoring component. Host manager also records the total energy consumption of the host computed using the power model defined as a function of CPU load.

Fields
id: u32rack_id: Option<u32>
Implementations

impl HostManager

pub fn new( rack_id: Option<u32>, cpu_total: u32, memory_total: u64, monitoring_id: u32, placement_store_id: u32, vm_api: Rc<RefCell<VmAPI>>, allow_vm_overcommit: bool, power_model: HostPowerModel, slav_metric: Box<dyn HostSLAVMetric>, ctx: SimulationContext, sim_config: Rc<SimulationConfig> ) -> Self

pub fn get_cpu_allocated(&self) -> f64

Returns the total amount of allocated vCPUs.

pub fn get_memory_allocated(&self) -> f64

Returns the total amount of allocated memory.

pub fn get_cpu_load(&self, time: f64) -> f64

Returns the current CPU load (used/total) by summing the resource consumption of all active VMs on this host.

pub fn get_memory_load(&self, time: f64) -> f64

Returns the current memory load (used/total) by summing the resource consumption of all active VMs on this host.

pub fn get_power(&self, cpu_load: f64) -> f64

Returns the current power consumption.

pub fn get_energy_consumed(&mut self, time: f64) -> f64

Returns the total energy consumption.

pub fn get_accumulated_slav(&mut self, time: f64) -> f64

Returns the total SLAV value.