Struct dslab_iaas::core::resource_pool::ResourcePoolState

pub struct ResourcePoolState { /* private fields */ }

Implementations

impl ResourcePoolState

pub fn new() -> Self

Creates empty resource pool state.

pub fn add_host( &mut self, id: u32, cpu_total: u32, memory_total: u64, cpu_available: u32, memory_available: u64, rack_id: Option<u32> )

Adds host to resource pool.

pub fn get_host_ids(&self) -> Vec<u32>

Returns IDs of all hosts.

pub fn get_hosts(&self) -> Values<'_, u32, HostInfo>

Returns an iterator over all hosts.

pub fn get_host(&self, host_id: u32) -> &HostInfo

Returns host info by its ID.

pub fn get_host_count(&self) -> u32

Returns the number of hosts.

pub fn can_allocate( &self, alloc: &Allocation, host_id: u32, allow_cpu_overcommit: bool ) -> AllocationVerdict

Checks if the specified allocation is currently possible on the specified host.

pub fn allocate(&mut self, alloc: &Allocation, host_id: u32)

Applies the specified application on the specified host.

pub fn release(&mut self, alloc: &Allocation, host_id: u32)

Removes the specified allocation on the specified host.

pub fn get_total_cpu(&self, host_id: u32) -> u32

Returns the total CPU capacity of the specified host.

pub fn get_total_memory(&self, host_id: u32) -> u64

Returns the total memory capacity of the specified host.

pub fn get_available_cpu(&self, host_id: u32) -> u32

Returns the amount of available vCPUs on the specified host.

pub fn get_available_memory(&self, host_id: u32) -> u64

Returns the amount of available memory on the specified host.

pub fn get_allocated_cpu(&self, host_id: u32) -> u32

Returns CPU capacity of the specified host currently in use by some VMs.

pub fn get_allocated_memory(&self, host_id: u32) -> u64

Returns memory capacity of the specified host currently in use by some VMs.

pub fn get_cpu_load(&self, host_id: u32) -> f64

Returns the CPU allocation rate (ratio of allocated to total resources) of the specified host

pub fn get_memory_load(&self, host_id: u32) -> f64

Returns the memory allocation rate (ratio of allocated to total resources) of the specified host