Struct dslab_iaas::core::vm::VirtualMachine

pub struct VirtualMachine {
    pub id: u32,
    pub cpu_usage: u32,
    pub memory_usage: u64,
    pub allocation_start_time: f64,
    /* private fields */
}

Represents virtual machine (VM).

VM is characterized by its ID, resource requirements (vCPUs and memory), start time, lifetime and load models. The latter model the actual resource utilization of VM in time, which may significantly differ from the VM’s resource requirements.

Fields
id: u32cpu_usage: u32memory_usage: u64allocation_start_time: f64
Implementations

impl VirtualMachine

pub fn new( id: u32, allocation_start_time: f64, lifetime: f64, resource_consumer: ResourceConsumer, sim_config: Rc<SimulationConfig> ) -> Self

Creates virtual machine with specified parameters.

pub fn lifetime(&self) -> f64

Returns VM lifetime (it is updated when VM is migrated).

pub fn start_time(&self) -> f64

Returns VM start time (it is updated when VM is migrated).

pub fn start_duration(&self) -> f64

Returns VM start duration (the value is taken from the simulation config).

pub fn stop_duration(&self) -> f64

Returns VM stop duration (the value is taken from the simulation config).

pub fn set_start_time(&mut self, time: f64)

Sets VM start time. Can be called multiple times due to VM migration.

pub fn set_lifetime(&mut self, lifetime: f64)

Changes VM lifetime. It is called only due to VM migration.

pub fn get_cpu_load(&self, time: f64) -> f64

Returns the current CPU load of VM by invoking the CPU load model.

pub fn get_memory_load(&self, time: f64) -> f64

Returns the current memory load of VM by invoking the memory load model.