Struct dslab_iaas::core::vm_api::VmAPI

pub struct VmAPI { /* private fields */ }

API to access information about virtual machines.

This component stores the information about all VMs in the system, including VM characteristics, current status and location, and provides an access to this information to other components. A user can also query VM API to obtain the needed VM information.

Implementations

impl VmAPI

pub fn new(ctx: SimulationContext) -> Self

Creates component.

pub fn get_id(&self) -> u32

Returns component ID.

pub fn register_new_vm(&mut self, vm: VirtualMachine)

Registers information about new VM. Called when VM is created via CloudSimulation.

pub fn get_vm_status(&self, vm_id: u32) -> VmStatus

Returns the status of specified VM.

Due to the asynchronous propagation of status updates, the returned status may be outdated.

pub fn get_vm(&self, vm_id: u32) -> Rc<RefCell<VirtualMachine>>

Returns the reference to VM information by VM ID.

pub fn get_vm_allocation(&self, vm_id: u32) -> Allocation

Returns resource allocation for specified VM.

pub fn find_host_by_vm(&self, vm_id: u32) -> Option<u32>

pub fn generate_vm_id(&mut self) -> u32

Generates new VM ID if user did not pass any.