Trait dslab_iaas::experiment::SimulationCallbacks

pub trait SimulationCallbacks: DynClone + Send {
    // Provided methods
    fn on_simulation_start(&mut self, _sim: &mut CloudSimulation) { ... }
    fn on_step(&mut self, _sim: &mut CloudSimulation) -> bool { ... }
    fn on_simulation_finish(
        &mut self,
        _sim: &mut CloudSimulation,
    ) -> IndexMap<String, String> { ... }
}

Trait for implementing custom callbacks for simulation runs within an experiment.

Provided Methods

fn on_simulation_start(&mut self, _sim: &mut CloudSimulation)

Runs before starting a simulation run.

fn on_step(&mut self, _sim: &mut CloudSimulation) -> bool

Runs on each step of a simulation run, returns false if the simulation must be stopped.

fn on_simulation_finish( &mut self, _sim: &mut CloudSimulation, ) -> IndexMap<String, String>

Runs upon the completion of a simulation run, returns results of this run.