Struct dslab_iaas::simulation::CloudSimulation

pub struct CloudSimulation { /* private fields */ }

Represents a simulation, provides methods for its configuration and execution.

It encapsulates all simulation components and provides convenient access to them for the user.

Implementations

impl CloudSimulation

pub fn new(sim: Simulation, sim_config: SimulationConfig) -> Self

Creates a simulation with specified config.

pub fn with_logger( sim: Simulation, sim_config: SimulationConfig, logger: Box<dyn Logger> ) -> Self

Creates a simulation with specified config.

pub fn add_host(&mut self, name: &str, cpu_total: u32, memory_total: u64) -> u32

Creates new host with specified name and resource capacity, and returns the host ID.

pub fn add_host_in_rack( &mut self, name: &str, cpu_total: u32, memory_total: u64, rack_id: u32 ) -> u32

Creates new host with specified name and resource capacity, and returns the host ID. Also associates the host with the specified rack.

pub fn add_scheduler( &mut self, name: &str, vm_placement_algorithm: VMPlacementAlgorithm ) -> u32

Creates new scheduler with specified name and VM placement algorithm, and returns the scheduler ID.

pub fn spawn_vm_now( &mut self, resource_consumer: ResourceConsumer, lifetime: f64, vm_id: Option<u32>, scheduler_id: u32 ) -> u32

Creates new VM with specified properties, registers it in VM API and immediately submits the allocation request to the specified scheduler. Returns VM ID.

pub fn spawn_vm_with_delay( &mut self, resource_consumer: ResourceConsumer, lifetime: f64, vm_id: Option<u32>, scheduler_id: u32, delay: f64 ) -> u32

Creates new VM with specified properties, registers it in VM API and submits the allocation request to the specified scheduler with the specified delay. Returns VM ID.

pub fn spawn_vm_on_host( &mut self, resource_consumer: ResourceConsumer, lifetime: f64, vm_id: Option<u32>, host_id: u32 ) -> u32

Creates new VM with specified properties and spawns it on the specified host bypassing the scheduling step. This is useful for creating the initial resource pool state.

pub fn begin_batch(&mut self)

Switches API to batch mode for building multi-VM requests. The subsequent invocations of spawn_vm_now will not spawn VM immediately but will add it to the batch. After the batch is completed it can be submitted using the spawn_batch method.

pub fn spawn_batch(&mut self) -> Vec<u32>

Spawns the current batch as a single multi-VM requests and disables the batch mode. Returns the IDs of spawned VMs.

pub fn migrate_vm_to_host(&mut self, vm_id: u32, target_host: u32)

Sends VM migration request to the specified target host.

pub fn build_custom_component<Component: 'static + CustomComponent>( &mut self, name: &str ) -> Rc<RefCell<Component>>

Creates custom component and adds it to the simulation.

pub fn spawn_vms_from_dataset( &mut self, default_scheduler_id: u32, dataset: &mut dyn DatasetReader )

Spawns all VMs from the given dataset.

The specified default scheduler is used for VM requests without scheduler information.

pub fn set_host_power_model(&mut self, host_power_model: HostPowerModel)

Overrides the used host power model.

Should be called before adding hosts to simulation.

pub fn set_slav_metric(&mut self, slav_metric: Box<dyn HostSLAVMetric>)

Overrides the used host-level SLAV metric.

Should be called before adding hosts to simulation.

pub fn monitoring(&self) -> Rc<RefCell<Monitoring>>

Returns the reference to monitoring component (provides actual host load).

pub fn vm_api(&self) -> Rc<RefCell<VmAPI>>

Returns the reference to VM API component (provides information about VMs).

pub fn context(&self) -> &SimulationContext

Returns the main simulation context.

pub fn steps(&mut self, step_count: u64) -> bool

Performs the specified number of steps through the simulation (see dslab-core docs).

pub fn step_for_duration(&mut self, time: f64)

Steps through the simulation with duration limit (see dslab-core docs).

pub fn step_until_time(&mut self, time: f64)

Steps through the simulation until the specified time (see dslab-core docs).

pub fn event_count(&self) -> u64

Returns the total number of created events.

pub fn current_time(&mut self) -> f64

Returns the current simulation time.

pub fn hosts(&self) -> BTreeMap<u32, Rc<RefCell<HostManager>>>

Returns the map with references to host managers.

pub fn host(&self, host_id: u32) -> Rc<RefCell<HostManager>>

Returns the reference to host manager (host energy consumption, allocated resources etc.).

pub fn host_by_name(&self, name: &str) -> Rc<RefCell<HostManager>>

Returns the reference to host manager (host energy consumption, allocated resources etc.).

pub fn scheduler(&self, scheduler_id: u32) -> Rc<RefCell<Scheduler>>

Returns the reference to scheduler.

pub fn scheduler_by_name(&self, name: &str) -> Rc<RefCell<Scheduler>>

Returns the reference to host scheduler.

pub fn vm(&self, vm_id: u32) -> Rc<RefCell<VirtualMachine>>

Returns the reference to VM information.

pub fn vm_status(&self, vm_id: u32) -> VmStatus

Returns the (possibly slightly outdated) status of specified VM via VM API.

pub fn vm_location(&self, vm_id: u32) -> Option<u32>

Returns the ID of host that runs the specified VM.

pub fn sim_config(&self) -> Rc<SimulationConfig>

Returns the simulation config.

pub fn lookup_id(&self, name: &str) -> Id

Returns the identifier of component by its name.

pub fn average_cpu_load(&mut self) -> f64

Returns the average CPU load across all hosts.

pub fn average_memory_load(&mut self) -> f64

Returns the average memory load across all hosts.

pub fn cpu_allocation_rate(&mut self) -> f64

Returns the current CPU allocation rate (% of overall CPU used).

pub fn memory_allocation_rate(&mut self) -> f64

Returns the current memory allocation rate (% of overall RAM used).

pub fn log_error(&mut self, log: String)

pub fn log_warn(&mut self, log: String)

pub fn log_info(&mut self, log: String)

pub fn log_debug(&mut self, log: String)

pub fn log_trace(&mut self, log: String)

pub fn save_log(&self, path: &str) -> Result<(), Error>