Module dslab_iaas::core

Modules

common

Common data structures.

config

energy_meter

Energy meter calculates the host energy consumption.

events

Standard simulation events.

host_manager

Host manager representing a physical machine.

load_model

Resource load models.

logger

monitoring

Service that provides information about current state of hosts.

placement_store

Component managing the authoritative copy of resource pool state.

resource_pool

Resource pool state.

scheduler

Component performing allocation of resources for new VMs.

slav_metric

Service-level agreement violation metrics.

vm

Representations of virtual machine and its status.

vm_api

Component that provides information about all VMs in the system.

vm_placement_algorithm

Virtual machine placement algorithms.

vm_placement_algorithms