Trait dslab_iaas::core::slav_metric::HostSLAVMetric

pub trait HostSLAVMetric: DynClone {
    // Required methods
    fn update(&mut self, time: f64, cpu_load: f64);
    fn value(&self) -> f64;
}

Trait for implementation of host-level SLA violation metric.

This metric measures the amount of SLA violation caused by the host overload, when it is not able to provide the full performance to hosted VMs.

Required Methods

fn update(&mut self, time: f64, cpu_load: f64)

Called whenever the host’s CPU load is changed to update the metric value.

fn value(&self) -> f64

Returns the current metric value.